fix tests import path
