+++ requirements.txt
@@ -1,2 +1,3 @@
 pandas
 yfinance
+pytest
